refactor(ui): extract turn-changes + diffs lists + helpers from session-turn
Slice 11b.1 Phase 2a. Bring `session-turn.tsx` under the AGENTS.md
500-line ceiling by extracting the two large render blocks plus the
pure helpers; the shell stays in place as the turn orchestrator (auto
scroll + status / error / thinking / divider + dispatch to Message and
AssistantParts).

New: packages/ui/src/components/session-turn-helpers.ts (135 lines)
  - record / unwrap        SDK error envelope unwrapping
  - same / list            shallow array helpers
  - HIDDEN_TOOLS           tool names suppressed from rendering
  - partState              visibility classifier
  - clean / heading        markdown heading extraction (thinking shimmer)

New: packages/ui/src/components/session-turn-changes-list.tsx (187 lines)
  TurnChangesList — file-by-file post-turn summary with chevron-toggle
  inline diff, open-file / show-in-folder actions, and the
  double-click-confirmed undo/redo button.

New: packages/ui/src/components/session-turn-diffs-list.tsx (133 lines)
  SessionTurnDiffsList — legacy sticky-header accordion for older
  sessions without turn-change metadata; lazy-mounts the diff body
  one frame after open to keep the transition smooth.

session-turn.tsx: 748 → 446 lines. Behaviour preserved — every memo,
signal, and effect stayed; only JSX render blocks moved out into
focused leaf components that take pre-resolved values as props.

tsgo --noEmit GREEN (app + ui). bun test (ui): 478 pass / 3 pre-existing
fail. bun test (app): 1050 pass / 0 fail.
